+++ bindings/nodejs/src/lib.rs
@@ -61,6 +61,7 @@ fn main(mut cx: ModuleContext) -> NeonResult<()> {
     cx.export_function("createWallet", wallet::create_wallet)?;
     cx.export_function("destroyWallet", wallet::destroy_wallet)?;
     cx.export_function("getClientFromWallet", wallet::get_client)?;
+    cx.export_function("getSecretManagerFromWallet", wallet::get_secret_manager)?;
     cx.export_function("listenWallet", wallet::listen_wallet)?;
 
     Ok(())