+++ bindings/python/src/wallet.rs
@@ -14,6 +14,7 @@ use tokio::sync::RwLock;
 use crate::{
     client::Client,
     error::{Error, Result},
+    SecretManager,
 };
 
 #[pyclass]
@@ -106,4 +107,27 @@ pub fn get_client_from_wallet(wallet: &Wallet) -> Result<Client> {
     })?;
 
     Ok(Client { client })
+}
+
+/// Get the secret manager from the wallet.
+#[pyfunction]
+pub fn get_secret_manager_from_wallet(wallet: &Wallet) -> Result<SecretManager> {
+    let secret_manager = crate::block_on(async {
+        wallet
+            .wallet
+            .read()
+            .await
+            .as_ref()
+            .map(|w| w.get_secret_manager().clone())
+            .ok_or_else(|| {
+                Error::from(
+                    serde_json::to_string(&Response::Panic("wallet got destroyed".into()))
+                        .expect("json to string error")
+                        .as_str(),
+                )
+            })
+    })?
+    .clone();
+
+    Ok(SecretManager { secret_manager })
 }