+++ sdk/CHANGELOG.md
@@ -37,6 +37,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Stronghold snapshot migration from v2 to v3;
 - `SecretManage::sign_evm`;
 - `Account::addresses_balance` method accepting addresses to get balance for;
+- `Wallet::get_secret_manager` method;
 
 ### Changed